+++ src/sampletones_application/layout/settings/__init__.py
@@ -4,6 +4,7 @@ from sampletones_application.layout.settings.audio import AudioSettingsLayout
 from sampletones_application.layout.settings.display import DisplaySettingsLayout
 from sampletones_application.layout.settings.export.export import ExportSettingsLayout
 from sampletones_application.layout.settings.keybindings import KeybindingsSettingsLayout
+from sampletones_application.layout.settings.nsf import NSFSettingsLayout
 from sampletones_application.layout.settings.render import RenderSettingsLayout
 
 
@@ -21,3 +22,4 @@ class SettingsLayout(BaseModel, extra="forbid", frozen=True):
     keybindings: KeybindingsSettingsLayout
     render: RenderSettingsLayout
     export: ExportSettingsLayout
+    nsf: NSFSettingsLayout